Fixed ActiveRecord attributes automatic typecast on attribute set.

+++ test/test_app/db/migrate/20090208151615_create_products.php
@@ -7,7 +7,7 @@ class CreateProduct extends ActiveRecord_Migration
     $t = $this->db->new_table('products');
     
     $t->add_column('string', 'name',  array('null' => false, 'limit' => 100));
-    $t->add_column('float',  'price', array('null' => false, 'signed' => false));
+    $t->add_column('double', 'price', array('null' => false, 'signed' => false));
     
     $t->add_timestamps();
     return $t->create();
